Move dialog classes to own file
