+++ osu.Game/Screens/Menu/MainMenu.cs
@@ -1,8 +1,6 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
-using System;
-using System.Collections.Generic;
 using System.Linq;
 using osuTK;
 using osuTK.Graphics;
@@ -10,7 +8,6 @@ using osu.Framework.Allocation;
 using osu.Framework.Audio.Track;
 using osu.Framework.Bindables;
 using osu.Framework.Graphics;
-using osu.Framework.Graphics.Sprites;
 using osu.Framework.Platform;
 using osu.Framework.Screens;
 using osu.Game.Configuration;
@@ -19,7 +16,6 @@ using osu.Game.Graphics.Containers;
 using osu.Game.IO;
 using osu.Game.Online.API;
 using osu.Game.Overlays;
-using osu.Game.Overlays.Dialog;
 using osu.Game.Screens.Backgrounds;
 using osu.Game.Screens.Edit;
 using osu.Game.Screens.Multi;
@@ -290,98 +286,6 @@ namespace osu.Game.Screens.Menu
 
             this.FadeOut(3000);
             return base.OnExiting(next);
-        }
-
-        private class ConfirmExitDialog : PopupDialog
-        {
-            public ConfirmExitDialog(Action confirm, Action cancel)
-            {
-                HeaderText = "Are you sure you want to exit?";
-                BodyText = "Last chance to back out.";
-
-                Icon = FontAwesome.Solid.ExclamationTriangle;
-
-                Buttons = new PopupDialogButton[]
-                {
-                    new PopupDialogOkButton
-                    {
-                        Text = @"Goodbye",
-                        Action = confirm
-                    },
-                    new PopupDialogCancelButton
-                    {
-                        Text = @"Just a little more",
-                        Action = cancel
-                    },
-                };
-            }
-        }
-
-        private class StorageErrorDialog : PopupDialog
-        {
-            [Resolved]
-            private DialogOverlay dialogOverlay { get; set; }
-
-            [Resolved]
-            private OsuGameBase osuGame { get; set; }
-
-            public StorageErrorDialog(OsuStorage storage, OsuStorageError error)
-            {
-                HeaderText = "osu! storage error";
-                Icon = FontAwesome.Solid.ExclamationTriangle;
-
-                var buttons = new List<PopupDialogButton>();
-
-                switch (error)
-                {
-                    case OsuStorageError.NotAccessible:
-                        BodyText = $"The specified osu! data location (\"{storage.CustomStoragePath}\") is not accessible. If it is on external storage, please reconnect the device and try again.";
-
-                        buttons.AddRange(new PopupDialogButton[]
-                        {
-                            new PopupDialogOkButton
-                            {
-                                Text = "Try again",
-                                Action = () =>
-                                {
-                                    if (!storage.TryChangeToCustomStorage(out var nextError))
-                                        dialogOverlay.Push(new StorageErrorDialog(storage, nextError));
-                                }
-                            },
-                            new PopupDialogOkButton
-                            {
-                                Text = "Reset to default location",
-                                Action = storage.ResetCustomStoragePath
-                            },
-                            new PopupDialogCancelButton
-                            {
-                                Text = "Use default location for this session",
-                            },
-                        });
-                        break;
-
-                    case OsuStorageError.AccessibleButEmpty:
-                        BodyText = $"The specified osu! data location (\"{storage.CustomStoragePath}\") is empty. If you have moved the files, please close osu! and move them back.";
-
-                        // Todo: Provide the option to search for the files similar to migration.
-                        buttons.AddRange(new PopupDialogButton[]
-                        {
-                            new PopupDialogCancelButton
-                            {
-                                Text = "Start fresh at specified location"
-                            },
-                            new PopupDialogOkButton
-                            {
-                                Text = "Reset to default location",
-                                Action = storage.ResetCustomStoragePath
-                            },
-                        });
-
-                        break;
-                }
-
-                Buttons = buttons;
-            }
         }
     }
 }
